==> 12.13score_card/MarTech_Challenge_20201230/1_Time_sliding_window.ipynb <==
import numpy as np  # linear algebra
import pandas as pd  #
from datetime import datetime, date, timedelta
from scipy.stats import skew  # for some statistics
from scipy.special import boxcox1p
from scipy.stats import boxcox_normmax
from sklearn.linear_model import ElasticNetCV, LassoCV, RidgeCV,Ridge,Lasso,ElasticNet
from sklearn.ensemble import GradientBoostingRegressor, RandomForestRegressor,RandomForestClassifier
from sklearn.feature_selection import mutual_info_regression
from sklearn.svm import SVR, LinearSVC
from sklearn.pipeline import make_pipeline,Pipeline
from sklearn.preprocessing import RobustScaler, LabelEncoder
from sklearn.model_selection import KFold, cross_val_score, train_test_split
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split
import os
import re
import seaborn as sns
import matplotlib.pyplot as plt
import time
from itertools import product
import datetime as dt
import calendar
import gc
import warnings
warnings.filterwarnings('ignore')

RANDOM_SEED = 42
# 引入中文字体
from matplotlib.font_manager import FontProperties
myfont = FontProperties(fname="/home/aistudio/NotoSansCJKsc-Light.otf", size=12)

PATH = '../data/'
train = pd.read_csv(PATH + 'train.csv')
# train = pd.read_csv('./data/data19383/train.csv', usecols=[2, 3, 4, 6, 7, 18])
# set index to ID to avoid droping it later
# 把测试集的id列作为索引，防止误删
test  = pd.read_csv(PATH + 'submission.csv').set_index('customer_id')

# 对于特别大的文件，我们需要做一些内存检查
mem_train = train.memory_usage(index=True).sum()
mem_test=test.memory_usage(index=True).sum()
print(u"训练集使用内容 "+ str(mem_train/ 1024**2)+" MB")
print(u"测试集使用内存 "+ str(mem_test/ 1024**2)+" MB")

# @from: https://www.kaggle.com/arjanso/reducing-dataframe-memory-size-by-65/code
# @liscense: Apache 2.0
# @author: weijian
def reduce_mem_usage(props):
    # 计算当前内存
    start_mem_usg = props.memory_usage().sum() / 1024 ** 2
    print("Memory usage of the dataframe is :", start_mem_usg, "MB")
    
    # 哪些列包含空值，空值用-999填充。why：因为np.nan当做float处理
    NAlist = []
    for col in props.columns:
        # 这里只过滤了objectd格式，如果你的代码中还包含其他类型，请一并过滤
        if (props[col].dtypes != object):
            
            # print("**************************")
            # print("columns: ", col)
            # print("dtype before", props[col].dtype)
            
            # 判断是否是int类型
            isInt = False
            mmax = props[col].max()
            mmin = props[col].min()
            
            # Integer does not support NA, therefore Na needs to be filled
            if not np.isfinite(props[col]).all():
                NAlist.append(col)
                props[col].fillna(-999, inplace=True) # 用-999填充
                
            # test if column can be converted to an integer
            asint = props[col].fillna(0).astype(np.int64)
            result = np.fabs(props[col] - asint)
            result = result.sum()
            if result < 0.01: # 绝对误差和小于0.01认为可以转换的，要根据task修改
                isInt = True
            
            # make interger / unsigned Integer datatypes
            if isInt:
                if mmin >= 0: # 最小值大于0，转换成无符号整型
                    if mmax <= 255:
                        props[col] = props[col].astype(np.uint8)
                    elif mmax <= 65535:
                        props[col] = props[col].astype(np.uint16)
                    elif mmax <= 4294967295:
                        props[col] = props[col].astype(np.uint32)
                    else:
                        props[col] = props[col].astype(np.uint64)
                else: # 转换成有符号整型
                    if mmin > np.iinfo(np.int8).min and mmax < np.iinfo(np.int8).max:
                        props[col] = props[col].astype(np.int8)
                    elif mmin > np.iinfo(np.int16).min and mmax < np.iinfo(np.int16).max:
                        props[col] = props[col].astype(np.int16)
                    elif mmin > np.iinfo(np.int32).min and mmax < np.iinfo(np.int32).max:
                        props[col] = props[col].astype(np.int32)
                    elif mmin > np.iinfo(np.int64).min and mmax < np.iinfo(np.int64).max:
                        props[col] = props[col].astype(np.int64)  
            else: # 注意：这里对于float都转换成float16，需要根据你的情况自己更改
                props[col] = props[col].astype(np.float16)
            
            # print("dtype after", props[col].dtype)
            # print("********************************")
    print("___MEMORY USAGE AFTER COMPLETION:___")
    mem_usg = props.memory_usage().sum() / 1024**2 
    print("Memory usage is: ",mem_usg," MB")
    print("This is ",100*mem_usg/start_mem_usg,"% of the initial size")
    return props, NAlist

# 处理id字段
train['order_detail_id'] = train['order_detail_id'].astype(np.uint32)
train['order_id'] = train['order_id'].astype(np.uint32)
train['customer_id'] = train['customer_id'].astype(np.uint32)
train['goods_id'] = train['goods_id'].astype(np.uint32)
train['goods_class_id'] = train['goods_class_id'].astype(np.uint32)
train['member_id'] = train['member_id'].astype(np.uint32)
# 处理状态字段，这里同时处理空值，将空值置为0
train['order_status'] = train['order_status'].astype(np.uint8)
train['goods_has_discount'] = train['goods_has_discount'].astype(np.uint8)
train["is_member_actived"].fillna(0, inplace=True)
train["is_member_actived"]=train["is_member_actived"].astype(np.int8)
train["member_status"].fillna(0, inplace=True)
train["member_status"]=train["member_status"].astype(np.int8)
train["customer_gender"].fillna(0, inplace=True)
train["customer_gender"]=train["customer_gender"].astype(np.int8)
train['is_customer_rate'] = train['is_customer_rate'].astype(np.uint8)
train['order_detail_status'] = train['order_detail_status'].astype(np.uint8)
# 处理日期
train['goods_list_time']=pd.to_datetime(train['goods_list_time'],format="%Y-%m-%d")
train['order_pay_time']=pd.to_datetime(train['order_pay_time'],format="%Y-%m-%d")
train['goods_delist_time']=pd.to_datetime(train['goods_delist_time'],format="%Y-%m-%d")
# 检查内存使用
mem_train = train.memory_usage(index=True).sum()
mem_test=test.memory_usage(index=True).sum()
print(u"训练集使用内容 "+ str(mem_train/ 1024**2)+" MB")
print(u"测试集使用内存 "+ str(mem_test/ 1024**2)+" MB")

train['customer_city_id'] = LabelEncoder().fit_transform(train['customer_city'].astype(str))
train['customer_province_id'] = LabelEncoder().fit_transform(train['customer_province'].astype(str))

# 纯时间序列滑窗
训练集和测试集构造思路：
- 训练集
	- 训练日期截止分别为2013.6.6、2013.6.20、2013.7.1、2013.7.15
	- 标签为训练截止日期开始的30天有没有下过单
	- 标签通过切片的形式构造
    - 把4个部分训练集拼起来
- 验证集
	- 验证数据到2013.8.1
	- 标签为2013.8.1开始的30天有没有下过单
	- 标签时间2013.8.1-2013.8.30
- 测试集
	- 测试数据到2013.8.31
	- 没有标签

train.head()

len(train['customer_id'].unique())

test.info()

test.head()

从上面的比对可以看出，这里要预测的是全体用户未来一个月的购买情况，训练集和测试集的`id`完全重合，没有需要特别处理的地方

## 构造时间滑窗特征

时间滑窗构造
- [已完成] 考虑到按周、按月的订单规律性变化，时间滑窗往往是7和30的倍数
- [已完成] 一般选取的特征是滑窗中的基本统计值，最大、最小、均值、中位数、求和等等
- [未完成] 节假日信息、影响周期的时间滑窗整合

# 将用户下单金额按天进行汇总
# df = train[train.order_status<101][train.order_pay_time>'2013-02-01']
df = train[train.order_pay_time>'2013-02-01']
df['date'] = pd.DatetimeIndex(df['order_pay_time']).date
df_payment = df[['customer_id','date','order_total_payment']]

len(df_payment['customer_id'].unique())

注意，成功交易的客户数量不等于全部客户数量，说明有相当一部分客户虽然下过单，但是没有成功的订单，那么这些客户自然应当算在训练集之外。
数据合并时，由于`test.csv`中，已经设置了默认0值，只需要和训练后的预测标签做一个`left join`就可以了

df_payment = df_payment.groupby(['date','customer_id']).agg({'order_total_payment': ['sum']})
df_payment.columns = ['day_total_payment']
df_payment.reset_index(inplace=True)

df_payment = df_payment.set_index(
    ["customer_id", "date"])[["day_total_payment"]].unstack(level=-1).fillna(0)
df_payment.columns = df_payment.columns.get_level_values(1)

df_payment.head()

df_goods = df[['customer_id','date','order_total_num']]
df_goods = df_goods.groupby(['date','customer_id']).agg({'order_total_num': ['sum']})
df_goods.columns = ['day_total_num']
df_goods.reset_index(inplace=True)
df_goods = df_goods.set_index(
    ["customer_id", "date"])[["day_total_num"]].unstack(level=-1).fillna(0)
df_goods.columns = df_goods.columns.get_level_values(1)

# 这是一个时间滑窗函数，获得dt之前minus天以来periods的dataframe，以便进一步计算
def get_timespan(df, dt, minus, periods, freq='D'):
    return df[pd.date_range(dt - timedelta(days=minus), periods=periods, freq=freq)]

1. 构造dataset这里有个取巧的地方，因为要预测的9月份除了开学季以外不是非常特殊的月份，因此主要考虑近期的因素，数据集的开始时间也是2月1日，尽量避免了双十一、元旦假期的影响，当然春节假期继续保留。同时，构造数据集的时候保留了customer_id，主要为了与其它特征做整合。
2. 通过一个函数整合付款金额和商品数量的时间滑窗，主要是因为分开做到时候合并占用内存更大，并且函数最后在返回值处做了内存优化，用时间代价尽可能避免内存溢出。

def prepare_dataset(df_payment, df_goods, t2018, is_train=True):
    X = {}
    # 整合用户id
    tmp = df_payment.reset_index()
    X['customer_id'] = tmp['customer_id']
    # 消费特征
    print('Preparing payment feature...')
    for i in [7,14,30,49,60,91,120]:
        tmp_1 = get_timespan(df_payment, t2018, i, i)
        # X['diff_%s_mean' % i] = tmp_1.diff(axis=1).mean(axis=1).values
        # X['mean_%s_decay' % i] = (tmp_1 * np.power(0.9, np.arange(i)[::-1])).sum(axis=1).values
        X['mean_%s' % i] = tmp_1.mean(axis=1).values
        X['median_%s' % i] = tmp_1.median(axis=1).values
        # X['min_%s' % i] = tmp_1.min(axis=1).values
        X['max_%s' % i] = tmp_1.max(axis=1).values
        # X['std_%s' % i] = tmp_1.std(axis=1).values
        X['sum_%s' % i] = tmp_1.sum(axis=1).values

        tmp_2 = get_timespan(df_payment, t2018 + timedelta(weeks=-1), i, i)
        # X['diff_%s_mean_2' % i] = tmp_2.diff(axis=1).mean(axis=1).values
        # X['mean_%s_decay_2' % i] = (tmp_2 * np.power(0.9, np.arange(i)[::-1])).sum(axis=1).values
        X['mean_%s_2' % i] = tmp_2.mean(axis=1).values
        X['median_%s_2' % i] = tmp_2.median(axis=1).values
        # X['min_%s_2' % i] = tmp_2.min(axis=1).values
        X['max_%s_2' % i] = tmp_2.max(axis=1).values
        # X['std_%s_2' % i] = tmp_2.std(axis=1).values

        tmp_3 = get_timespan(df_payment, t2018, i, i)
        X['has_sales_days_in_last_%s' % i] = (tmp_3 != 0).sum(axis=1).values
        X['last_has_sales_day_in_last_%s' % i] = i - ((tmp_3 != 0) * np.arange(i)).max(axis=1).values
        X['first_has_sales_day_in_last_%s' % i] = ((tmp_3 != 0) * np.arange(i, 0, -1)).max(axis=1).values

    # 对此处进行微调，主要考虑近期因素
    for i in range(1, 4):
        X['day_%s_2018' % i] = get_timespan(df_payment, t2018, i*30, 30).sum(axis=1).values

    for i in range(7):
        X['mean_4_dow{}_2013'.format(i)] = get_timespan(df_payment, t2018, 56-i*2, 4, freq='14D').mean(axis=1).values
        # X['mean_20_dow{}_2013'.format(i)] = get_timespan(df, t2018, 140-i, 20, freq='7D').mean(axis=1).values
        X['mean_20_dow{}_2013'.format(i)] = get_timespan(df_payment, t2018, 140-i*2, 10, freq='14D').mean(axis=1).values
    
    # 商品数量特征，这里故意把时间和消费特征错开，提高时间滑窗的覆盖面
    print('Preparing num feature...')
    for i in [3,21,35,49,70,84,105]:
            tmp_1 = get_timespan(df_goods, t2018, i, i)
            # X['goods_diff_%s_mean' % i] = tmp_1.diff(axis=1).mean(axis=1).values
            # X['goods_mean_%s_decay' % i] = (tmp_1 * np.power(0.9, np.arange(i)[::-1])).sum(axis=1).values
            X['goods_mean_%s' % i] = tmp_1.mean(axis=1).values
            X['goods_median_%s' % i] = tmp_1.median(axis=1).values
            # X['goods_min_%s' % i] = tmp_1.min(axis=1).values
            X['goods_max_%s' % i] = tmp_1.max(axis=1).values
            # X['goods_std_%s' % i] = tmp_1.std(axis=1).values
            X['goods_sum_%s' % i] = tmp_1.sum(axis=1).values
    
            tmp_2 = get_timespan(df_goods, t2018 + timedelta(weeks=-1), i, i)
            # X['goods_diff_%s_mean_2' % i] = tmp_2.diff(axis=1).mean(axis=1).values
            # X['goods_mean_%s_decay_2' % i] = (tmp_2 * np.power(0.9, np.arange(i)[::-1])).sum(axis=1).values
            X['goods_mean_%s_2' % i] = tmp_2.mean(axis=1).values
            X['goods_median_%s_2' % i] = tmp_2.median(axis=1).values
            # X['goods_min_%s_2' % i] = tmp_2.min(axis=1).values
            X['goods_max_%s_2' % i] = tmp_2.max(axis=1).values
            X['goods_sum_%s_2' % i] = tmp_2.sum(axis=1).values
    
            tmp_3 = get_timespan(df_goods, t2018, i, i)
            X['goods_has_sales_days_in_last_%s' % i] = (tmp_3 > 0).sum(axis=1).values
            X['goods_last_has_sales_day_in_last_%s' % i] = i - ((tmp_3 > 0) * np.arange(i)).max(axis=1).values
            X['goods_first_has_sales_day_in_last_%s' % i] = ((tmp_3 > 0) * np.arange(i, 0, -1)).max(axis=1).values


    # 对此处进行微调，主要考虑近期因素
    for i in range(1, 4):
        X['goods_day_%s_2018' % i] = get_timespan(df_goods, t2018, i*28, 28).sum(axis=1).values

    for i in range(7):
        X['goods_mean_4_dow{}_2013'.format(i)] = get_timespan(df_goods, t2018, 56-i*2, 4, freq='14D').mean(axis=1).values
        # X['mean_20_dow{}_2013'.format(i   = get_timespan(df, t2018, 140-i, 20, freq='7D').mean(axis=1).values
        X['goods_mean_20_dow{}_2013'.format(i)] = get_timespan(df_goods, t2018, 140-i*2, 10, freq='14D').mean(axis=1).values

    X = pd.DataFrame(X)
    
    reduce_mem_usage(X)
    
    if is_train:
        # 这样转换之后，打标签直接用numpy切片就可以了
        # 当然这里前提是确认付款总额没有负数的问题
        y = df_goods[pd.date_range(t2018, periods=30)].max(axis=1).values
        y[y > 0] = 1
        return X, y
    return X

## 生成滑窗数据
此处对内存占用要求非常高，尽量分步操作避免crash，即使如此，内存溢出仍然频繁出现，通过下面的办法进一步压缩数据集大小

### 生成训练集
对此处进行优化，把训练集不曾涉及到的新增用户剔除，也就是说，模型研究的是已有用户的购买行为变化趋势。如果一个用户在8月中旬才第一次下单，那么就不会被放到训练集/验证集中，只会出现在测试集中。

内存限制的情况下，训练集和验证集时间的选取会比较纠结，既要考虑月份，也要考虑星期。

def make_slides(train,df_part,begin,end,column):
    # 将用户下单金额按天进行汇总
    # df = train[train.order_status <= 6][train.order_pay_time > '2013-02-01']
    df = train[train.order_pay_time > begin][train.order_pay_time < end]
    df = pd.merge(df,df_part,how='inner')
    df['date'] = pd.DatetimeIndex(df['order_pay_time']).date
    df = df[['customer_id', 'date', column]]
    df = df.groupby(['date', 'customer_id']).agg({column: ['sum']})
    df.columns = ['day_' + column]
    df.reset_index(inplace=True)
    df = df.set_index(
        ["customer_id", "date"])[['day_' + column]].unstack(level=-1).fillna(0)
    df.columns = df.columns.get_level_values(1)
    return df

#### part1
2013.06.06之前的用户，在2013.06.06后30天的时间滑窗

#get_timespan(df_goods, date(2013, 8, 1), 120-7*2, 10, freq='14D')

# df_0606 pred_start-5个月, pred_start
# make_slides跨度范围时 pred_start-5个月，pred_start+30
# prepare_dataset 传入的时pred_start

df_0606 = train[train.order_pay_time > '2013-01-01'][train.order_pay_time <= '2013-06-06'][['customer_id']]
# 删除重复行
df_0606 = df_0606.drop_duplicates(['customer_id'])
df_part1_partment = make_slides(train,df_0606,'2013-01-01','2013-07-06','order_total_payment')
df_part1_goods = make_slides(train,df_0606,'2013-01-01','2013-07-06','order_total_num')
X_part1, y_part1 = prepare_dataset(df_part1_partment, df_part1_goods, date(2013, 6, 6))

X_part1.to_pickle('work/X_part1.pkl')
np.save("work/y_part1.npy", y_part1)

del df_0606
del df_part1_partment
del df_part1_goods
del X_part1
del y_part1
gc.collect()

#### part2
2013.06.20之前的用户，在2013.06.20后30天的时间滑窗

df_0620 = train[train.order_pay_time > '2013-01-15'][train.order_pay_time <= '2013-06-20'][['customer_id']]
# 删除重复行
df_0620 = df_0620.drop_duplicates(['customer_id'])
df_part2_partment = make_slides(train,df_0620,'2013-01-15','2013-07-20','order_total_payment')
df_part2_goods = make_slides(train,df_0620,'2013-01-15','2013-07-20','order_total_num')
X_part2, y_part2 = prepare_dataset(df_part2_partment, df_part2_goods, date(2013, 6, 20))

X_part2.to_pickle('work/X_part2.pkl')
np.save("work/y_part2.npy", y_part2)

del df_0620
del df_part2_partment
del df_part2_goods
del X_part2
del y_part2
gc.collect()

#### part3
2013.07.01之前的用户，在2013.07.01后30天的时间滑窗

df_0701 = train[train.order_pay_time > '2013-01-26'][train.order_pay_time <= '2013-07-01'][['customer_id']]
# 删除重复行
df_0701 = df_0701.drop_duplicates(['customer_id'])
df_part3_partment = make_slides(train,df_0701,'2013-01-26','2013-07-31','order_total_payment')
df_part3_goods = make_slides(train,df_0701,'2013-01-26','2013-07-31','order_total_num')
X_part3, y_part3 = prepare_dataset(df_part3_partment, df_part3_goods, date(2013, 7, 1))

X_part3.to_pickle('work/X_part3.pkl')
np.save("work/y_part3.npy", y_part3)

del df_0701
del df_part3_partment
del df_part3_goods
del X_part3
del y_part3
gc.collect()

#### part4
2013.07.15之前的用户，在2013.07.15后30天的时间滑窗

df_0715 = train[train.order_pay_time > '2013-02-10'][train.order_pay_time <= '2013-07-15'][['customer_id']]
# 删除重复行
df_0715 = df_0715.drop_duplicates(['customer_id'])
df_part4_partment = make_slides(train,df_0715,'2013-02-10','2013-08-16','order_total_payment')
df_part4_goods = make_slides(train,df_0715,'2013-02-10','2013-08-16','order_total_num')
X_part4, y_part4 = prepare_dataset(df_part4_partment, df_part4_goods, date(2013, 7, 15))

X_part4.to_pickle('work/X_part4.pkl')
np.save("work/y_part4.npy", y_part4)

del df_0715
del df_part4_partment
del df_part4_goods
del X_part4
del y_part4
gc.collect()

print("Preparing Train dataset...")
X_l, y_l = [], []
X_part1 = pd.read_pickle('work/X_part1.pkl')
y_part1 = np.load('work/y_part1.npy')
X_l.append(X_part1)
y_l.append(y_part1)
X_part2 = pd.read_pickle('work/X_part2.pkl')
y_part2 = np.load('work/y_part2.npy')
X_l.append(X_part2)
y_l.append(y_part2)
X_part3 = pd.read_pickle('work/X_part3.pkl')
y_part3 = np.load('work/y_part3.npy')
X_l.append(X_part3)
y_l.append(y_part3)
X_part4 = pd.read_pickle('work/X_part4.pkl')
y_part4 = np.load('work/y_part4.npy')
X_l.append(X_part4)
y_l.append(y_part4)
X_train = pd.concat(X_l, axis=0)
y_train = np.concatenate(y_l, axis=0)
print('done!')
del X_l
del y_l
del X_part1
del y_part1
del X_part2
del y_part2
del X_part3
del y_part3
del X_part4
del y_part4
gc.collect()
X_train.to_pickle('work/X_train.pkl')
np.save("work/y_train.npy", y_train)

### 生成验证集
2013.08.01之前的用户，在2013.08.01后30天的时间滑窗

print("Preparing Validation dataset...")
df_0801 = train[train.order_pay_time > '2013-03-01'][train.order_pay_time <= '2013-08-01'][['customer_id']]
# 删除重复行
df_0801 = df_0801.drop_duplicates(['customer_id'])
df_part5_partment = make_slides(train,df_0801,'2013-03-01','2013-08-31','order_total_payment')
df_part5_goods = make_slides(train,df_0801,'2013-03-01','2013-08-31','order_total_num')
X_val, y_val = prepare_dataset(df_part5_partment, df_part5_goods, date(2013, 8, 1))

X_val.to_pickle('work/X_val.pkl')
np.save("work/y_val.npy", y_val)

del df_0801
del df_part5_partment
del df_part5_goods
gc.collect()

### 生成测试集

df_0901 = train[train.order_pay_time > '2013-03-01'][train.order_pay_time < '2013-09-01'][['customer_id']]
# 删除重复行
df_0901 = df_0901.drop_duplicates(['customer_id'])
df_test_partment = make_slides(train,df_0901,'2013-04-01','2013-09-01','order_total_payment')
df_test_goods = make_slides(train,df_0901,'2013-04-01','2013-09-01','order_total_num')
X_test = prepare_dataset(df_test_partment, df_test_goods, date(2013, 9, 1),is_train=False)

X_test.to_pickle('work/X_test.pkl')

至此，整合了用户id的消费时间滑窗就完整了

### 小结：
- 纯时间序列滑窗也可以直接用模型开始训练，但是效果并不太好。
考虑到还有其它重要的数据维度，此时需要构建新的交叉特征。如用户X商品，用户X会员，用户X城市，用户X省份等，然后与时间滑窗数据整合。

- 纯时间滑窗还有个问题，面对当前这个非常稀疏的数据集，浪费了大量的内存空间。

#### 用户基本特征

把地域、会员这些基本信息整合到训练数据中

df_customer = train[['member_id','customer_id','customer_gender','customer_province_id','customer_city_id','is_member_actived']]
# 删除重复行
df_customer = df_customer.drop_duplicates(['customer_id'])
df_customer.head()

X_train = pd.merge(X_train,df_customer,how='left')
X_val = pd.merge(X_val,df_customer,how='left')
X_test = pd.merge(X_test,df_customer,how='left')

X_train.to_pickle('X_train.pkl')
X_val.to_pickle('X_val.pkl')
X_test.to_pickle('X_test.pkl')
