==> 12.13score_card/MarTech_Challenge_20201230/3_merge_slidingwindow_user.ipynb <==
import pickle

# 查看fea.pkl 和。。。
with open('../data/product_fea.pkl', 'rb') as f:
    product = pickle.load(f)
with open('../data/user_fea.pkl', 'rb') as f:
    user = pickle.load(f)

product.columns

user.columns

PATH = '../data/'
with open(PATH + 'sliding_windows/X_train.pkl', 'rb') as f:
    x_train = pickle.load(f)
with open(PATH + 'sliding_windows/X_val.pkl', 'rb') as f:
    x_val = pickle.load(f)
with open(PATH + 'sliding_windows/X_test.pkl', 'rb') as f:
    x_test = pickle.load(f)

import numpy as np
y_val = np.load('./work/y_val.npy')
y_train = np.load('./work/y_train.npy')

x_train.shape, x_val.shape

y_train.shape, y_val.shape

x_train['label'] = y_train
x_val['label'] = y_val

# 融合用户统计特征
x_train_user = x_train.merge(user, on='customer_id')
x_val_user = x_val.merge(user, on='customer_id')
x_test_user = x_test.merge(user, on='customer_id')

y_train_user = x_train_user.label
y_val_user = x_val_user.label

x_train_user.shape, x_val_user.shape

y_train_user.shape, y_val_user.shape

##### 保存

np.save("work/y_train_user.npy", y_train_user)
np.save("work/y_val_user.npy", y_val_user)

y_train_user.drop(columns='label', inplace=True)
y_val_user.drop(columns='label', inplace=True)

x_train_user.to_pickle(PATH + 'sliding_windows/x_train_user.pkl')
x_val_user.to_pickle(PATH + 'sliding_windows/x_val_user.pkl')
x_test_user.to_pickle(PATH + 'sliding_windows/x_test_user.pkl')
